pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "media-duplicate-checker"
version = "1.0.0"
description = "Cross-platform video and audio duplicate finder with GUI and CLI"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "VoxHash Technologies", email = "contact@voxhash.dev"}
]
keywords = ["video", "audio", "duplicate", "deduplication", "media", "files"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "mutagen>=1.47.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "imagehash>=4.3.1",
    "PyQt6>=6.6.0",
    "rapidfuzz>=3.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "coverage>=7.3.0",
]

[project.scripts]
media-duplicate-checker = "media_checker.app:main"

[tool.setuptools]
packages = ["media_checker", "bookmark_checker"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "PyQt6.*",
    "cv2.*",
    "mutagen.*",
    "PIL.*",
    "imagehash.*",
    "rapidfuzz.*",
    "bs4.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-q --cov=media_checker --cov=bookmark_checker --cov-report=term-missing --cov-report=xml --cov-fail-under=85"

[tool.coverage.run]
source = ["media_checker", "bookmark_checker"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/__main__.py",
    "*/app.py",
    "*/ui/*",
    "*/i18n/translations.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

